+++ history.md
@@ -1,4 +1,11 @@
 # 版本变更
+## 0.5.0 (2019-01-07)
+* [EDIT] yyt 配置文件 改为 `yyt.config.js`, `config.js` 不再支持
+* [ADD] 新增 `yyt --proxy` 配置项
+* [ADD] 新增 `yyt --headless` 配置项
+* [ADD] 新增 `config.proxy` 配置项
+* [ADD] 新增 `config.headless` 配置项
+
 ## 0.4.0 (2018-12-24)
 * [EDIT] 默认配置文件 改为 `yyt.config.js`
 * [ADD] 新增 `yyt path/to/config` 方式